backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
stripe==7.8.0
bcrypt==4.1.2
email-validator==2.1.0
Pillow==10.1.0
pandas==2.1.4
numpy==1.24.3
aiofiles==23.2.1
jinja2==3.1.2
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0